requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyTelegramBotAPI==3.7.5
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.2
Werkzeug==1.0.1